+++ src/file_monitor.rs
@@ -222,14 +222,14 @@ pub trait FileMonitor {
     /// Return true if the ioctl should be fully emulated. If so the result
     /// is stored in the last parameter.
     /// Only called during recording.
-    fn emulate_ioctl(&mut self, _t: &RecordTask, _r: &mut u64) -> bool {
+    fn emulate_ioctl(&mut self, _t: &RecordTask, _r: &mut usize) -> bool {
         false
     }
 
     /// Return true if the fcntl should should be fully emulated. If so the
     /// result is stored in the last parameter.
     /// Only called during recording.
-    fn emulate_fcntl(&self, _t: &RecordTask, _r: &mut u64) -> bool {
+    fn emulate_fcntl(&self, _t: &RecordTask, _r: &mut usize) -> bool {
         false
     }
